// frontend/src/client/types.gen.ts
export type ProblemRead = {
    number: number;
    render_mode: string;
    body_latex?: (string | null);
    image_path?: (string | null);
    choices?: (Array<{
    [key: string]: unknown;
}> | null);
};